Inline getExtensionStatus
